99_Archive/11_get-publ-data/99_archive/get-publ-data_v2/requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
webdriver-manager==4.0.1

